+++ app/components/bali/side_menu/list/component.rb
@@ -6,13 +6,14 @@ module Bali
       class Component < ApplicationViewComponent
         renders_many :items,
                      lambda { |href: nil, name: nil, icon: nil,
-                               authorized: true, disabled: false, **options|
+                               authorized: true, disabled: false, target: nil, **options|
                        Item::Component.new(
                          name: name,
                          href: href,
                          icon: icon,
                          authorized: authorized,
                          disabled: disabled,
+                         target: target,
                          current_path: @current_path,
                          group_behavior: @group_behavior,
                          **options